+++ .github/workflows/pr-tests.yml
@@ -146,10 +146,14 @@ jobs:
         run: rustup target add riscv32imac-unknown-none-elf --toolchain nightly-2024-08-01-x86_64-unknown-linux-gnu
       - name: Install test dependencies
         run: sudo apt-get update && sudo apt-get install -y binutils-riscv64-unknown-elf lld
-      - name: Run examples with tagged powdr lib
-        run: cargo run --profile pr-tests --example hello_world && cargo run --profile pr-tests --example sqrt_with_publics && cargo run --profile pr-tests --example fibonacci
-      - name: Run examples with PR's powdr lib
+      - name: Run examples that cargo accepts as examples
+        run: cargo run --profile pr-tests --example hello_world && cargo run --profile pr-tests --example sqrt_with_publics
+      - name: Run crate example serialized_inputs with the given branch
         run: cd powdr-test/examples/serialized_inputs && cargo run --profile pr-tests
+      - name: Run crate example fibonacci with the given branch
+        run: cd powdr-test/examples/fibonacci && cargo run --profile pr-tests
+      - name: Run crate example fibonacci with the latest tag
+        run: cd examples/fibonacci && cargo run --profile pr-tests
 
   test_estark_polygon:
     needs: build
