+++ powdr-test/examples/fibonacci/src/main.rs
@@ -3,7 +3,7 @@ use powdr::Session;
 fn main() {
     env_logger::init();
 
-    let n = 11;
+    let n = 22;
     let mut session = Session::builder()
         .guest_path("./guest")
         .out_path("powdr-target")
@@ -14,19 +14,16 @@ fn main() {
     session.run();
 
     let r: u32 = session.stdout();
-    assert_eq!(r, 89);
+    assert_eq!(r, 17711);
 
     session.prove();
 
     let r: u32 = session.stdout();
-    assert_eq!(r, 89);
+    assert_eq!(r, 17711);
 
     let publics = session.publics();
     assert_eq!(
         publics,
-        [
-            555233681, 1854640251, 3298928347, 2857173302, 2660189392, 1608424695, 543896544,
-            3870154745
-        ]
+        [1743086182, 3059275111, 427636475, 2461890199, 2369471556, 1082999957, 4188715485, 3633080053]
     );
 }
